+++ project1/AUTHORS
@@ -1,0 +1,8 @@
+# Authors
+
+This project wouldn't have been possible without the contributions of the following people:
+
+## Core Developers
+- Stef Neys (stef.neys@student.kuleuven.be)
+- Ula Kocur (ula.kocur@student.kuleuven.be)
+- Stef Eversdijk (stef.eversdijk@student.kuleuven.be)
