+++ project1/tests/README.org
@@ -1,0 +1,10 @@
+* Expected energies
+
+|        |      E(HF) |      E(MP2) |
+|--------+------------+-------------|
+| CH_4   |  -40.19867 |  -40.362681 |
+| H_2O   |  -76.02679 |  -76.230759 |
+| C_2H_2 |  -76.82577 |  -77.085873 |
+| HCN    |  -92.88295 |  -93.172222 |
+| H_3COH | -115.04842 | -115.391507 |
+| CO_2   | -187.65071 | -188.139385 |
